refactor: order와 Ordertable 패키지 분리.
