+++ order/src/main/java/kitchenpos/ordertable/dto/OrderTableRequest.java
@@ -1,6 +1,6 @@
-package kitchenpos.order.dto;
+package kitchenpos.ordertable.dto;
 
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 
 public class OrderTableRequest {
