+++ order/src/main/java/kitchenpos/ordertable/dto/OrderTableResponse.java
@@ -1,7 +1,7 @@
-package kitchenpos.order.dto;
+package kitchenpos.ordertable.dto;
 
 import java.util.Objects;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 
 public class OrderTableResponse {
