+++ order/src/main/java/kitchenpos/ordertable/validator/OrderTableValidator.java
@@ -1,10 +1,10 @@
-package kitchenpos.order.validator;
+package kitchenpos.ordertable.validator;
 
 import java.util.Objects;
 import java.util.Optional;
 import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 import org.springframework.stereotype.Component;
 
 @Component
