+++ order/src/main/java/kitchenpos/table/validator/TableGroupValidator.java
@@ -3,9 +3,9 @@ package kitchenpos.table.validator;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.order.repository.OrderRepository;
-import kitchenpos.order.repository.OrderTableRepository;
+import kitchenpos.ordertable.repository.OrderTableRepository;
 import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.table.domain.TableGroup;
 import kitchenpos.table.dto.TableGroupRequest;
 import org.springframework.stereotype.Component;
