+++ order/src/test/java/kitchenpos/order/validator/OrderTableValidatorTest.java
@@ -6,7 +6,8 @@ import static org.mockito.ArgumentMatchers.eq;
 import java.util.Optional;
 import kitchenpos.order.OrderFixture;
 import kitchenpos.order.repository.OrderRepository;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.validator.OrderTableValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
